refactor(step10): Controller에 존재하는 비즈니스 로직을 Service로 이전
